+++ MyWallet/Models/CategoryModel.cs
@@ -74,10 +74,23 @@ namespace MyWallet.Models
             objDAL.UpdateCategory(this);
         }
 
-        public void Remove(int id)
+        public bool Remove(int id)
         {
-            string sql = $"DELETE FROM CATEGORY WHERE ID={id}";
-            new DAL().ExecuteSQLCommand(sql);
+            string sql = $"SELECT id from transaction where category_id = {id}";
+
+            DAL objDAL = new DAL();
+            DataTable dt = objDAL.RetrieveDataTable(sql);
+
+            if (dt.Rows.Count > 0)
+            {
+                return false;
+            }
+            else
+            {
+                sql = $"DELETE FROM CATEGORY WHERE ID={id}";
+                objDAL.ExecuteSQLCommand(sql);
+                return true;
+            }
         }
 
         public CategoryModel LoadRegister(int? id)
